rename to atoms/chem_comp and bonds/chem_comp
