+++ src/mol-io/writer/ligand-encoder.ts
@@ -7,8 +7,8 @@
 import { StringBuilder } from '../../mol-util';
 import Writer from './writer';
 import { Encoder, Category, Field } from './cif/encoder';
-import { ComponentAtom } from '../../mol-model-formats/structure/property/bonds/atom';
-import { ComponentBond } from '../../mol-model-formats/structure/property/bonds/comp';
+import { ComponentAtom } from '../../mol-model-formats/structure/property/atoms/chem_comp';
+import { ComponentBond } from '../../mol-model-formats/structure/property/bonds/chem_comp';
 
 interface Atom {
     Cartn_x: number,
