+++ src/mol-model-formats/structure/property/atoms/chem_comp.ts
