+++ src/mol-model-formats/structure/property/bonds/chem_comp.ts
