+++ src/servers/model/server/query.ts
@@ -22,13 +22,13 @@ import CifField = CifWriter.Field
 import { splitCamelCase } from '../../../mol-util/string';
 import { Encoder } from '../../../mol-io/writer/cif/encoder';
 import { Encoding } from './api';
-import { ComponentBond } from '../../../mol-model-formats/structure/property/bonds/comp';
+import { ComponentBond } from '../../../mol-model-formats/structure/property/bonds/chem_comp';
 import { SdfWriter } from '../../../mol-io/writer/sdf';
 import { MolWriter } from '../../../mol-io/writer/mol';
 import { Mol2Writer } from '../../../mol-io/writer/mol2';
 import { MolEncoder } from '../../../mol-io/writer/mol/encoder';
 import { Mol2Encoder } from '../../../mol-io/writer/mol2/encoder';
-import { ComponentAtom } from '../../../mol-model-formats/structure/property/bonds/atom';
+import { ComponentAtom } from '../../../mol-model-formats/structure/property/atoms/chem_comp';
 
 export interface Stats {
     structure: StructureWrapper,
